Improve C++ tooling.
This commit adds support for compile_commands.json to Bazel. This allows
us to use Coc and ccls in Vim. Very nice. Compared to vim-go and gopls,
this was a certified PITA to get working.

I might refactor parts of this out into a dedicated repository at some
point. Now I'm just glad it works.

This is what AoC is really about. Learning new things and finally
getting that Vim/C++ setup working.

For convenience, it also adds a Makefile with shortcuts for formatting,
linting etc.